+++ server/service/src/processors/transfer/invoice/delete_inbound_invoice.rs
